requirements.txt
langchain
langchain_community
langchain-ollama
FastAPI
Uvicorn 
googlesearch-python
langdetect
textblob

